# public/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "iTaxCix API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api/v1"
        }
    ],
    "paths": {
...
        "/admin/role-permission/create": {
            "post": {
                "tags": [
                    "RolePermission"
                ],
                "summary": "Asignar permiso a rol",
                "description": "Asigna un permiso a un rol.",
                "operationId": "createRolePermission",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RolePermissionCreateRequestDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Asignación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Creado"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/RolePermissionResponseDTO"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Errores de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Petición inválida"
                                        },
                                        "error": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Campos requeridos faltantes"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No autorizado"
                                        },
                                        "error": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Token inválido o expirado"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/role-permission/update": {
            "put": {
                "tags": [
                    "RolePermission"
                ],
                "summary": "Actualizar asignación de permiso a rol",
                "description": "Actualiza la asignación de un permiso a un rol.",
                "operationId": "updateRolePermission",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RolePermissionUpdateRequestDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actualización exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "OK"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/RolePermissionResponseDTO"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/role-permission/delete": {
            "delete": {
                "tags": [
                    "RolePermission"
                ],
                "summary": "Eliminar asignación de permiso a rol",
                "description": "Elimina la asignación de un permiso a un rol.",
                "operationId": "deleteRolePermission",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RolePermissionDeleteRequestDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Eliminación exitosa (sin contenido)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/role-permission/list": {
            "get": {
                "tags": [
                    "RolePermission"
                ],
                "summary": "Listar asignaciones de permisos a roles",
                "description": "Lista todas las asignaciones de permisos a roles.",
                "operationId": "listRolePermissions",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Cantidad de elementos por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "OK"
                                        },
                                        "data": {
                                            "properties": {
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/RolePermissionResponseDTO"
                                                    }
                                                },
                                                "meta": {
                                                    "properties": {
                                                        "total": {
                                                            "type": "integer",
                                                            "example": 50
                                                        },
                                                        "perPage": {
                                                            "type": "integer",
                                                            "example": 10
                                                        },
                                                        "currentPage": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "lastPage": {
                                                            "type": "integer",
                                                            "example": 5
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "RolePermissionCreateRequestDTO": {
                "description": "Datos para crear asignación de permiso a rol",
                "properties": {
                    "roleId": {
                        "description": "ID del rol",
                        "type": "integer",
                        "example": 2
                    },
                    "permissionId": {
                        "description": "ID del permiso",
                        "type": "integer",
                        "example": 3
                    },
                    "active": {
                        "description": "Si la asignación está activa",
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
            "RolePermissionDeleteRequestDTO": {
                "description": "Datos para eliminar asignación de permiso a rol",
                "properties": {
                    "id": {
                        "description": "ID de la asignación",
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "RolePermissionResponseDTO": {
                "description": "Datos de respuesta de asignación de permiso a rol",
                "properties": {
                    "id": {
                        "description": "ID de la asignación",
                        "type": "integer",
                        "example": 1
                    },
                    "roleId": {
                        "description": "ID del rol",
                        "type": "integer",
                        "example": 2
                    },
                    "permissionId": {
                        "description": "ID del permiso",
                        "type": "integer",
                        "example": 3
                    },
                    "active": {
                        "description": "Si la asignación está activa",
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
            "RolePermissionUpdateRequestDTO": {
                "description": "Datos para actualizar asignación de permiso a rol",
                "properties": {
                    "id": {
                        "description": "ID de la asignación",
                        "type": "integer",
                        "example": 1
                    },
                    "roleId": {
                        "description": "ID del rol",
                        "type": "integer",
                        "example": 2
                    },
                    "permissionId": {
                        "description": "ID del permiso",
                        "type": "integer",
                        "example": 3
                    },
                    "active": {
                        "description": "Si la asignación está activa",
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}